Move options to networks
